requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cleanup==8.1.0
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
